fix: sql 수정

+++ step17-board-backend/src/main/java/com/spring/dto/BoardReactionReq.java
@@ -12,6 +12,6 @@ public class BoardReactionReq {
   private int id;
   private int bno;
   private String type;
-  private int mid;
+  private Long mid;
 
 }
